last map change in tavern
I moved the key area  name="Thieves03:3"  back to the guest room. It was a good idea to move it onto the entrence, like the other keys.
But in the tavern it is different. Here the entrance mapchange is deactivated together with the rest area. So you can't enter the tavern, if yo get out of the room (which is possible, but only by death)